Year-2/Digital Image Processing/requirements.txt
customtkinter
matplotlib
pillow
numpy
scipy
scikit-image